+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Added gtm event when a product is added or removed from the comparison (vtex:compareProducts)
 ## [0.17.0] - 2022-06-15
 
 ### Added